requirements.txt
Django==1.7.1
Pillow==2.6.1
Unidecode==0.04.16
django-filer==0.9.8
django-mptt==0.6.1
django-polymorphic==0.6
easy-thumbnails==2.2
